refactor: move suppliers routes, schemas, and service to suppliers/
- api/routes/suppliers.py → suppliers/routes.py
- schemas/supplier.py → suppliers/schemas.py
- services/supplier_service.py → suppliers/service.py
- Update imports in routes and test patches
