+++ app/suppliers/service.py
