+++ tests/functional/test_supplier_endpoints.py
@@ -58,7 +58,7 @@ async def test_list_suppliers_returns_200(app, client):
     membership = _mock_membership()
 
     with _supplier_dependency_overrides(app, membership):
-        with patch("app.services.supplier_service.list_suppliers", return_value=[]):
+        with patch("app.suppliers.service.list_suppliers", return_value=[]):
             response = await client.get(
                 "/api/suppliers",
                 headers={"Authorization": "Bearer fake"},
@@ -73,7 +73,7 @@ async def test_create_supplier_returns_201(app, client):
 
     with _supplier_dependency_overrides(app, membership):
         with patch(
-            "app.services.supplier_service.create_supplier", return_value=new_supplier
+            "app.suppliers.service.create_supplier", return_value=new_supplier
         ):
             response = await client.post(
                 "/api/suppliers",
@@ -101,8 +101,8 @@ async def test_delete_works_for_owner(app, client):
     supplier = _mock_supplier(org_id=membership.org_id)
 
     with _supplier_dependency_overrides(app, membership):
-        with patch("app.services.supplier_service.get_supplier", return_value=supplier):
-            with patch("app.services.supplier_service.delete_supplier"):
+        with patch("app.suppliers.service.get_supplier", return_value=supplier):
+            with patch("app.suppliers.service.delete_supplier"):
                 response = await client.delete(
                     f"/api/suppliers/{supplier.id}",
                     headers={"Authorization": "Bearer fake"},
@@ -114,7 +114,7 @@ async def test_get_nonexistent_returns_404(app, client):
     membership = _mock_membership()
 
     with _supplier_dependency_overrides(app, membership):
-        with patch("app.services.supplier_service.get_supplier", return_value=None):
+        with patch("app.suppliers.service.get_supplier", return_value=None):
             response = await client.get(
                 f"/api/suppliers/{uuid.uuid4()}",
                 headers={"Authorization": "Bearer fake"},
